assembly 整合,清除重名问题
