+++ src/backend/ci/core/boot-assembly/build.gradle
@@ -33,15 +33,19 @@ dependencies {
 
     compile project(":core:artifactory:biz-artifactory-store")
     compile project(":core:artifactory:biz-artifactory-sample")
-    compile project(":core:dispatch:biz-dispatch")
+    compile project(":core:dispatch:biz-dispatch-docker")
     compile project(":core:environment:biz-environment-sample")
     compile project(":core:log:biz-log")
+    compile project(":core:image:biz-image")
+    compile project(":core:notify:biz-notify-blueking")
     compile project(":core:plugin:biz-plugin")
     compile project(":core:process:biz-process-sample")
     compile project(":core:project:biz-project-sample")
+    compile project(":core:quality:biz-quality")
     compile project(":core:repository:biz-repository-sample")
     compile project(":core:ticket:biz-ticket-sample")
     compile project(":core:store:biz-store-sample")
+    compile project(":core:websocket:biz-websocket")
 }
 
 apply from: "$rootDir/task_spring_boot_package.gradle"
