+++ src/backend/ci/core/common/common-pipeline/common-pipeline-api/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.common.pipeline.AutoConfiguration
+com.tencent.devops.common.pipeline.CommonPipelineAutoConfiguration
