+++ src/backend/ci/core/dispatch/boot-dispatch/build.gradle
@@ -25,7 +25,7 @@
  */
 
 dependencies {
-    compile project(":core:dispatch:biz-dispatch")
+    compile project(":core:dispatch:biz-dispatch-docker")
     compile project(":core:common:common-auth:common-auth-sample")
 }
