+++ src/backend/ci/core/plugin/codecc-plugin/biz-codecc/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.plugin.codecc.init.MQConfiguration
+com.tencent.devops.plugin.codecc.init.CodeCCMQConfiguration
