+++ src/backend/ci/core/plugin/codecc-plugin/common-codecc/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.plugin.codecc.AutoConfiguration
+com.tencent.devops.plugin.codecc.CodeCCAutoConfiguration
