+++ src/backend/ci/core/process/biz-process/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineBuildDetailService.kt
@@ -62,7 +62,7 @@ class PipelineBuildDetailService @Autowired constructor(
     private val redisOperation: RedisOperation,
     private val pipelineEventDispatcher: PipelineEventDispatcher,
     private val webSocketDispatcher: WebSocketDispatcher,
-    private val websocketService: WebsocketService,
+    private val pipelineWebsocketService: PipelineWebsocketService,
     private val pipelineBuildDao: PipelineBuildDao
 ) {
 
@@ -134,7 +134,7 @@ class PipelineBuildDetailService @Autowired constructor(
         val pipelineBuildInfo = pipelineBuildDao.getBuildInfo(dslContext, buildId) ?: return
         logger.info("dispatch pipelineDetailChangeEvent, buildId: $buildId")
         webSocketDispatcher.dispatch(
-                websocketService.buildDetailMessage(pipelineBuildInfo.buildId, pipelineBuildInfo.projectId, pipelineBuildInfo.pipelineId, pipelineBuildInfo.startUser)
+                pipelineWebsocketService.buildDetailMessage(pipelineBuildInfo.buildId, pipelineBuildInfo.projectId, pipelineBuildInfo.pipelineId, pipelineBuildInfo.startUser)
         )
 //        pipelineEventDispatcher.dispatch(
 //            PipelineStatusChangeEvent(
