+++ src/backend/ci/core/quality/common-quality/src/main/kotlin/com/tencent/devops/quality/QualityAutoConfiguration.kt
@@ -24,29 +24,20 @@
  * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package com.tencent.devops.process.api
+package com.tencent.devops.quality
 
-import com.tencent.devops.common.api.pojo.Result
-import com.tencent.devops.common.web.RestResource
-import com.tencent.devops.process.api.service.ServiceScmResource
-import com.tencent.devops.process.engine.pojo.event.commit.GithubWebhookEvent
-import com.tencent.devops.process.engine.service.PipelineBuildWebhookService
-import com.tencent.devops.process.engine.webhook.CodeWebhookEventDispatcher
-import com.tencent.devops.process.pojo.code.WebhookCommit
-import com.tencent.devops.process.pojo.code.github.GithubWebhook
-import org.springframework.amqp.rabbit.core.RabbitTemplate
-import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.boot.autoconfigure.AutoConfigureOrder
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.core.Ordered
 
-@RestResource
-class ServiceScmResourceImpl @Autowired constructor(
-    private val pipelineBuildService: PipelineBuildWebhookService,
-    private val rabbitTemplate: RabbitTemplate
-) : ServiceScmResource {
-    override fun webHookCodeGithubCommit(webhook: GithubWebhook): Result<Boolean> {
-        return Result(CodeWebhookEventDispatcher.dispatchGithubEvent(rabbitTemplate, GithubWebhookEvent(githubWebhook = webhook)))
-    }
+@Configuration
+@AutoConfigureOrder(Ordered.LOWEST_PRECEDENCE)
+class QualityAutoConfiguration {
 
-    override fun webhookCommit(projectId: String, webhookCommit: WebhookCommit): Result<String> {
-        return Result(pipelineBuildService.webhookCommitTriggerPipelineBuild(projectId, webhookCommit))
-    }
+    @Bean
+    fun qualityGateInElementBizPlugin() = QualityGateInElementBizPlugin()
+
+    @Bean
+    fun qualityGateOutElementBizPlugin() = QualityGateOutElementBizPlugin()
 }
