+++ src/backend/ci/core/quality/common-quality/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,2 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-com.tencent.devops.quality.AutoConfiguration
+com.tencent.devops.quality.QualityAutoConfiguration
