+++ src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/service/github/GithubService.kt
@@ -37,7 +37,7 @@ import com.tencent.devops.common.api.util.OkhttpUtils
 import com.tencent.devops.common.api.util.ShaUtils
 import com.tencent.devops.common.client.Client
 import com.tencent.devops.common.service.utils.RetryUtils
-import com.tencent.devops.process.api.service.ServiceScmResource
+import com.tencent.devops.process.api.service.ServiceScmWebhookResource
 import com.tencent.devops.process.pojo.code.github.GithubWebhook
 import com.tencent.devops.repository.pojo.AuthorizeResult
 import com.tencent.devops.repository.pojo.GithubCheckRuns
@@ -80,7 +80,7 @@ class GithubService @Autowired constructor(
                 return
             }
 
-            client.get(ServiceScmResource::class)
+            client.get(ServiceScmWebhookResource::class)
                 .webHookCodeGithubCommit(GithubWebhook(event, guid, removePrefixSignature, body))
         } catch (t: Throwable) {
             logger.info("Github webhook exception", t)
