+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/configuration/StoreMQConfiguration.kt
@@ -22,7 +22,7 @@ import org.springframework.context.annotation.Configuration
  * 流水线构建扩展配置
  */
 @Configuration
-class MQConfiguration {
+class StoreMQConfiguration {
 
     @Bean
     fun rabbitAdmin(connectionFactory: ConnectionFactory): RabbitAdmin {
