+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/dao/template/StoreTemplateDao.kt
@@ -45,7 +45,7 @@ import org.jooq.impl.DSL
 import org.springframework.stereotype.Repository
 
 @Repository
-class TemplateDao {
+class StoreTemplateDao {
 
     fun list(
         dslContext: DSLContext,
