+++ src/backend/ci/core/store/biz-store/src/main/kotlin/com/tencent/devops/store/service/websocket/StoreWebsocketService.kt
@@ -41,7 +41,7 @@ import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
 
 @Service
-class WebsocketService @Autowired constructor(
+class StoreWebsocketService @Autowired constructor(
     val redisOperation: RedisOperation,
     val objectMapper: ObjectMapper,
     val dslContext: DSLContext,
