+++ src/backend/ci/core/websocket/biz-websocket/src/main/kotlin/com/tencent/devops/websocket/configuration/WebSocketConfig.kt
@@ -44,7 +44,7 @@ class WebSocketConfig @Autowired constructor(
     private val bkHandshake: BKHandshakeInterceptor
 ) : AbstractWebSocketMessageBrokerConfigurer() {
 
-    @Value("\${thread.min}")
+    @Value("\${thread.min:8}")
     private val min: Int = 8
 
     companion object {
